Player-Script angepasst
-Player.cs ist ab jetzt zuständig für Punktevergabe
-Score wird in Szene angezeigt
-Player-Sprite wird je nach Bewegungsrichtung gespiegelt

+++ Potty Jump/Assets/MyData/Scripts/LevelGenerator.cs
@@ -57,7 +57,7 @@ public class LevelGenerator : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        int playerPoints = player.GetComponent<Player>().Points;
+        float playerPoints = player.GetComponent<Player>().Points;
 
         //Debug.Log(playerPoints);
